chore: use images to render SpectrumMeter part-1
Use images to render spectrum meter
 - a background image
 - a foreground image
Typically the backgound image is a "faint" version desaturated
of the foreground image
Neither of these images are required, if only one is used then
that is rendered.

This is also a cheap and easy way to implement different colour
gradients for the spectrum meter simply set the foreground image to the
required image.

Part 2 is scaling the images to fit.