chore: VUMeter + Spectrum enhancements
Enhance VUMeters and Spectrum Meter for Joggler, PiGrid and Touch skins
TODO fix other skins:

Introduce a caching to avoid rescaling and cropping VU meter and Spectrum images
for repeatedly.
Scan for VUMeter and Spectrum graphics images and pre-process them for faster caching
for future runs.
Cached images are resized and possibly cropped versions of the original
and are stored in ~/.jivelite/userpath/cache as bmp files.
The cache directory is created if necessary when images are pre-processed.
For now all disk cached images are loaded into an in memory-cache
 - a future change will only load require images into memory

Add a VU Meter selection menu item which allows selection of the VU Meter
graphics to use. Multiple VU Meter graphics may be selected, in which case
a different VU Meter is used when the track changes.
Spectrum graphics menu item is TBD