chore: integrate visualiser changes into HD skins
HD skins now have 3 nowplaying screens,
 Artwork and Text
 Spectrum Analyser
 VU Meter

In all 3 screens the album art is display prominently
as seemed to have been the original intention.
Spectrum and VU Meters are rendered to the right,
below Album title and above progress bar.

Progress bar and controls were moved further down
a bit.

To make the space for the vui