chore: Joggler skin wallpaper refactor
Make changes so that the wallpaper is integrated into all screens.
One side effect is that the now playing screen widgets are flat.
Previous appearance is almost restored by choising the black
wallpaper as background.

Full restoration can be achieved by setting BLACK_BLACKGROUND to
true in JogglerSkinApplet.lua

I hope to make this configurable via a menu item in a future change.