chore: hide volume controls if fixed volume
Almost works correctly, if Fixed Volume setting is changed
this isn't reflected in the NowPlaying view immediately,
but on the next view refresh.
AFAICT fixing this seemingly simple bug appears non-trivial
So the behaviour is deemed good enough for now, as it is unlikely
that users will be constantly changing this setting.