fix: crash on resize of older vfd VU meters
some/older vfd VU meters do not have images for
right channel and re-use the left channel images.

The resize function releases the source image,
unconditionally.
This is incorrect, it should only release the source
image if a resize was performed.
Otherwise on resize if left and right channels are
using the same source image, it if freed twice,
resulting in a crash.