chore: feature: change visualiser graphics on timer
All spectrum meter and vumeters objects have the same
shape.
It is possible to re-initialise the live object with new parameters
object whilst the NowPlaying view is in progress
This makes it appear that a new meter is being
displayed.

This change add a timer configuration and UI support to set
the timer.

The meter objects countdown after each frame is rendered and
initiate a refresh.

Inline resizing of visualiser resources is unhandled,
in these chase visualisation will show blank screens.