fix: move configuration items from envvars to settings
Move
 - save as png
 - save resized image
 - workspace

from environment variables to settings with UI support.
Make migration smooth by using the environment variables
if the settings are nil.

Note workspace UI needs more UI work for full user clarity on pCP.
For now establish the persistent storage root and attempt to coax
the user to use that as the root path for the workspace.
TBD.