fix: behaviour when rezied images are deleted
images queued for resize are permanently in the
queue, this allows us to handle multiple requests
for the same resize operation correctly, which
is return the result of previously enqueued
operation.

When resize images are deleted using the UI,
this queue is no longer valid.
Clear the queue if the conditions allow it,
namely a resize is not pending, i.e.
clearing the queue is thread safe.