bugfix: visualiser selection
remove subtle interplay between UI and visualiser selection
Instead forcing enablement of at least one VU or Spectrum
meter - allow none to be selected.

If a now playing view with a visulaiser is then selected,
a "default" visualiser is selected preferentially one
that does not require resizing.

for UI consistency when a visualiser is selected force
visualiser change in the visImage module.
This yields expected behaviour when current visualisers are deselected
and a different set of visualisers are enabled.