chore: Add option to startup Jivelite in NowPlaying screen
defaults to off, selectable in the Now Playing Menu
when on, switches to NowPlaying view after a configured value in settings
Default the value to 5000 milliseconds
This is more likely to work on slow platforms
No UI to change the configured value for now.
Advanced users can edit NowPlaying settings and change the value manually.