chore: Joggler: scaling: add source images
Add source images from which the scaled versions are created
to match the display resolution.
The source images were copied from the HDGridKsin.
Symlinks to the HDGrid skin tree could have been used,
however for simplicity and reduction in dependencies,
images have been copied.
The images in IconsResized we renamed as
 *_touch.png
from
 *_remote.png

Only images which are scaled have been copied.
This means that images that are composed from
multiple images may use sets of images which
are in the same directory.
Rational - scope rediction: scaling code does not
have to also decide which images to scale,
that responsibilty lies solely within the skin.