chore: annotate track and album fields
Annotate track with disc.tracknumber retreived
from audio metatdata
Annotate album with year retrieved from
audio metadata

The annotations are not guaranteed because the annotation
is performed after audio metadata is retrieved  from
the LMS server.

Ideally this data would be present in player status,
that appears to be more involved.